power/qcom: Ensure correct power-on reason on AC wake from hibernate
When a device exits from hibernate due to an AC connection, the AP
firmware must receive the correct power-on reason (CBLPWR). This change
introduces a mechanism to facilitate this.

A new Kconfig option, , is added.
When enabled, the power sequencing code checks if the wake-up source
from hibernate is AC. If so, it initiates the power-on sequence with
an AC power-on event. This ensures the AP correctly identifies the
reason for the wake-up.

This feature is implemented generically, allowing individual boards to
provide their own  function to detect the wake source.

The implementation details are as follows:
1.  **Declaration of board-specific function:** A new function prototype
    `int board_check_hibernate_wake_source_ac(void);`, is added to
    `include/power/qcom.h`. Boards enabling
    `CONFIG_PLATFORM_EC_HIBERNATE_EXIT_AC_WAKE` are expected to provide
    their own implementation of this function to detect if AC is the
    wake source.
2.  **Integration in power sequencing:** In `power/qcom_exp.c`,
    within the `power_chipset_init` function, a conditional block is
    added under `CONFIG_PLATFORM_EC_HIBERNATE_EXIT_AC_WAKE`. If the
    `EC_RESET_FLAG_HIBERNATE` flag is set and
    `board_check_hibernate_wake_source_ac()` returns true,
    `auto_power_on` is set to 0 and `ac_on` is set to 1. This directs
    the power state machine to use `POWER_ON_BY_AC_ON` as the power-on
    reason, which corresponds to the desired CBLPWR event for the AP
    firmware.

LOW_COVERAGE_REASON=b:422080009 tracking test code
BUG=b:431796208
TEST=Tested on a bluey device with this feature enabled.

Change-Id: I82c8c682f70dfb786959879ee3f65f13cb75cbb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/7253569
Auto-Submit: Jayvik Desai <jayvik@google.com>
Commit-Queue: Jayvik Desai <jayvik@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Tested-by: Jayvik Desai <jayvik@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/7537526
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Jon Murphy <jpmurphy@google.com>
Tested-by: Jon Murphy <jpmurphy@google.com>