gsctool: add vendor command for reading owner's config
Owner's configuration is stored on Opentitan chips in a dedicated INFO
page. This patch introduces a vendor command which would allow to read
the contents of the owner's configuration page in multiple read
transactions, each transaction retrieving 48 bytes, starting at the
increasing offsets into the page. The retrieved data is saved in a
binary file.

BUG=none
TEST=with the corresponding change in the GSC code read the config
     from the prod chip, and verified its validity using
     opentitantool:

     gsctool -Q /tmp/ot.cfg.bin
     opentitantool  ownership  config -i /tmp/owner.cfg  --inform binary

Change-Id: I5903ce1b86e202a088eaf189ca27da8c13da8f31
Signed-off-by: Vadim Bendebury <vbendeb@google.com>