zephyr: test: Add RISC-V IRQ stack guard validation test
This change introduces a new test case to validate the functionality
of the Physical Memory Protection (PMP) for the Interrupt Request
(IRQ) stack guard on RISC-V architectures.

The test (`pmp_irq_stack`) attempts to write to the memory region
designated as the IRQ stack guard. It verifies that this action
triggers a CPU exception, as expected, confirming that the stack
guard is correctly protected by the PMP. This ensures that
potential stack overflows into the guard region are caught.

BUG=b:469067565
TEST=./test/run_device_tests.py -b sanok -t pmp_irq_stack --zephyr;
     flash on ET171 board; ztest run-all

Change-Id: Iecb5ae87fc96555c5e859931aafa41396eb22d89