zephyr: shim: Define CONFIG_FINGERPRINT_MCU via fingerprint board config
Map the `CONFIG_BOARD_FINGERPRINT` Kconfig to the internal
`CONFIG_FINGERPRINT_MCU` macro within the Zephyr shim.

This change prevents `get_feature_flags1()` from reporting the
`EC_FEATURE_EXEC_IN_RAM` flag when a fingerprint board is in use.
Disabling this flag is necessary for the host-side `flashrom` tool to
function correctly, as it requires the ability to trigger a reboot to
the RO firmware before initiating a firmware update

BUG=b:445467149
TEST=Verified 'ectool inventory' no longer shows "Execute code in RAM"
     on helipilot.

Change-Id: I85a517e766f20588f4f3f25c7a63a1de9336a667